backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
redis[hiredis]==5.0.4
python-jose[cryptography]==3.3.0
argon2-cffi==23.1.0
httpx==0.27.0
websockets==12.0
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
apscheduler==3.10.4
pandas==2.2.2
numpy==1.26.4
ta==0.11.0
openai==1.30.1
yfinance==0.2.38
alembic==1.13.1
python-dotenv==1.0.1
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.2.2
aiofiles==23.2.1
greenlet==3.0.3
